+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 ## [unreleased]
 ### Changed
 - Validate and persist OIDC token in cookie for report and overview endpoints
+- Condensed summary endpoint not protected by auth for now
 
 ## [3.8]
 ### Changed 
 - In order to support token login via Google, validate `id_token` instead of `access_token` in overview and report endpoints.
-- Condensed summary endpoint not protected by auth for now
 
 ## [3.7]
 ### Added